Conflict between two pieces of the same color
